Move non-service stuff into terminal contrib
